--- a/container/go.mod
+++ b/container/go.mod
@@ -15,10 +15,10 @@ require (
 	github.com/containerd/platforms v0.2.1
 	github.com/docker/docker v28.2.2+incompatible
 	github.com/docker/go-connections v0.5.0
-	github.com/docker/go-sdk/client v0.1.0-alpha004
-	github.com/docker/go-sdk/config v0.1.0-alpha004
-	github.com/docker/go-sdk/image v0.1.0-alpha004
-	github.com/docker/go-sdk/network v0.1.0-alpha004
+	github.com/docker/go-sdk/client v0.1.0-alpha005
+	github.com/docker/go-sdk/config v0.1.0-alpha005
+	github.com/docker/go-sdk/image v0.1.0-alpha005
+	github.com/docker/go-sdk/network v0.1.0-alpha005
 	github.com/stretchr/testify v1.10.0
 	golang.org/x/sys v0.33.0
 )
@@ -31,7 +31,7 @@ require (
 	github.com/containerd/log v0.1.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/distribution/reference v0.6.0 // indirect
-	github.com/docker/go-sdk/context v0.1.0-alpha004 // indirect
+	github.com/docker/go-sdk/context v0.1.0-alpha005 // indirect
 	github.com/docker/go-units v0.5.0 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect