--- a/image/options.go
+++ b/image/options.go
@@ -2,6 +2,7 @@ package image
 
 import (
 	"github.com/docker/docker/api/types/image"
+	"github.com/docker/docker/client"
 )
 
 // PullOption is a function that configures the pull options.
@@ -24,6 +25,30 @@ func WithPullClient(pullClient ImagePullClient) PullOption {
 func WithPullOptions(imagePullOptions image.PullOptions) PullOption {
 	return func(opts *pullOptions) error {
 		opts.pullOptions = imagePullOptions
+		return nil
+	}
+}
+
+// SaveOption is a function that configures the save options.
+type SaveOption func(*saveOptions) error
+
+type saveOptions struct {
+	saveClient  ImageSaveClient
+	saveOptions []client.ImageSaveOption
+}
+
+// WithSaveClient sets the save client used to save the image.
+func WithSaveClient(saveClient ImageSaveClient) SaveOption {
+	return func(opts *saveOptions) error {
+		opts.saveClient = saveClient
+		return nil
+	}
+}
+
+// WithSaveOptions sets the save options used to save the image.
+func WithSaveOptions(options ...client.ImageSaveOption) SaveOption {
+	return func(opts *saveOptions) error {
+		opts.saveOptions = options
 		return nil
 	}
 }