--- a/image/testdata/Dockerfile
+++ b/image/testdata/Dockerfile
@@ -1,0 +1,4 @@
+ARG tag=latest
+   FROM nginx:${tag}
+COPY . .
+CMD ["pwd"]