--- a/image/testdata/Dockerfile.multistage
+++ b/image/testdata/Dockerfile.multistage
@@ -1,0 +1,4 @@
+FROM nginx:a as builderA
+FROM nginx:b as builderB
+FROM nginx:c as builderC
+FROM scratch