--- a/image/testdata/Dockerfile.multistage.singleBuildArgs
+++ b/image/testdata/Dockerfile.multistage.singleBuildArgs
@@ -1,0 +1,5 @@
+ARG BASE_IMAGE=scratch
+FROM nginx:a as builderA
+FROM nginx:b as builderB
+FROM nginx:c as builderC
+FROM ${BASE_IMAGE}