… image/testdata/Dockerfile.multistage.multiBuildArgs