@mdelapenya 2025-06-23
I don't know what goes on with the benchmarks here. I'm running them locally, and they pass the 1.5 threshold. Also, re-running the failing benchmark is not deterministic, as sometimes some benchmarks pass the threshold, and sometimes they don't.

**PR**
```
➜ network (images-work) ✗ go test ./... -cpu=4 -benchmem -run=^$ -bench .
goos: darwin
goarch: arm64
pkg: github.com/docker/go-sdk/network
cpu: Apple M1 Pro
BenchmarkNetworkOperations/create-network-4                    5         212629558 ns/op           34408 B/op        223 allocs/op
BenchmarkNetworkOperations/create-network-with-options-4                       6         178716062 ns/op           42022 B/op        231 allocs/op
BenchmarkNetworkOperations/inspect-network-4                             6547477               184.6 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkOperations/inspect-network-with-cache-4                  6214660               174.3 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkOperations/inspect-network-without-cache-4                    30          34917107 ns/op           14553 B/op        140 allocs/op
BenchmarkNetworkOperations/terminate-network-4                                 5         214841225 ns/op           20934 B/op        221 allocs/op
BenchmarkNetworkOperations/create-and-terminate-network-4                      5         210764308 ns/op           39604 B/op        227 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-creation-4                       8         140440875 ns/op           23837 B/op        230 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-inspection-with-cache-4            3631564               324.1 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-inspection-with-no-cache-4             108           9942579 ns/op           14772 B/op        142 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-termination-4                            8         141269536 ns/op           28743 B/op        231 allocs/op
PASS
ok      github.com/docker/go-sdk/network        28.149s
```

**main**
```
➜ network (main) ✗ go test ./... -cpu=4 -benchmem -run=^$ -bench .
goos: darwin
goarch: arm64
pkg: github.com/docker/go-sdk/network
cpu: Apple M1 Pro
BenchmarkNetworkOperations/create-network-4                    5         204247900 ns/op           27672 B/op        222 allocs/op
BenchmarkNetworkOperations/create-network-with-options-4                       6         177843979 ns/op           39126 B/op        229 allocs/op
BenchmarkNetworkOperations/inspect-network-4                             6482508               217.8 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkOperations/inspect-network-with-cache-4                  5868187               179.0 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkOperations/inspect-network-without-cache-4                    32          35303930 ns/op           14230 B/op        140 allocs/op
BenchmarkNetworkOperations/terminate-network-4                                 5         226339475 ns/op           27721 B/op        222 allocs/op
BenchmarkNetworkOperations/create-and-terminate-network-4                      5         212503300 ns/op           37950 B/op        226 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-creation-4                       8         135885208 ns/op           23565 B/op        230 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-inspection-with-cache-4            3792421               318.8 ns/op            32 B/op          1 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-inspection-with-no-cache-4             102           9963881 ns/op           14472 B/op        140 allocs/op
BenchmarkNetworkConcurrent/concurrent-network-termination-4                            8         138521312 ns/op           23449 B/op        230 allocs/op
PASS
ok      github.com/docker/go-sdk/network        25.430s
```

**Comparing main with the PR**
```diff
- BenchmarkNetworkOperations/create-network-with-options-4                       6         177843979 ns/op           39126 B/op        229 allocs/op
+ BenchmarkNetworkOperations/create-network-with-options-4                       6         178716062 ns/op           42022 B/op        231 allocs/op

- BenchmarkNetworkOperations/create-and-terminate-network-4                      5         212503300 ns/op           37950 B/op        226 allocs/op
+ BenchmarkNetworkOperations/create-and-terminate-network-4                      5         210764308 ns/op           39604 B/op        227 allocs/op

- BenchmarkNetworkConcurrent/concurrent-network-creation-4                       8         135885208 ns/op           23565 B/op        230 allocs/op
+ BenchmarkNetworkConcurrent/concurrent-network-creation-4                       8         140440875 ns/op           23837 B/op        230 allocs/op
```